UI: Split ViewBindingAdapter class
